media-source/MediaSource-addSourceBuffer_infinite-calling.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Infinite calling MediaSource#addSourceBuffer</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = '/media/white.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';
async_test(function(t) {
    var mediaSource = new MediaSource();
    mediaSource.addEventListener('sourceopen', function(e){
        var sourceBuffer = null;
        t.step(function() {
            assert_throws(
                {name : 'QuotaExceededError'},
                function() {
                    while(true) {
                        sourceBuffer = mediaSource.addSourceBuffer('video/webm; codecs="vorbis,vp8"');
                    }I
                },
                'Create infinite sourceBuffers'
            );
        });
        t.done();
    }, false);
    video.src = window.URL.createObjectURL(mediaSource);
}, "MediaSource#addSourceBuffer : If the user agent can't handle any more SourceBuffer objects then throw a QUOTA_EXCEEDED_ERR exception and abort these steps");
</script>
</body>
</html>
